+++ source/Energinet.DataHub.MarketParticipant.Application/Commands/GridArea/GetGridAreaAuditLogEntriesCommand.cs
@@ -13,17 +13,9 @@
 // limitations under the License.
 
 using System;
+using MediatR;
 
 namespace Energinet.DataHub.MarketParticipant.Application.Commands.GridArea
 {
-    public sealed record GridAreaOverviewItemDto(
-        Guid Id,
-        string Code,
-        string Name,
-        string PriceAreaCode,
-        DateTimeOffset ValidFrom,
-        DateTimeOffset? ValidTo,
-        string? ActorNumber,
-        string? ActorName,
-        DateTimeOffset? FullFlexDate);
+    public sealed record GetGridAreaAuditLogEntriesCommand(Guid GridAreaId) : IRequest<GetGridAreaAuditLogEntriesResponse>;
 }