+++ source/Energinet.DataHub.MarketParticipant.Application/Commands/GridArea/GetGridAreaAuditLogEntriesResponse.cs
@@ -12,18 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
+using System.Collections.Generic;
 
 namespace Energinet.DataHub.MarketParticipant.Application.Commands.GridArea
 {
-    public sealed record GridAreaOverviewItemDto(
-        Guid Id,
-        string Code,
-        string Name,
-        string PriceAreaCode,
-        DateTimeOffset ValidFrom,
-        DateTimeOffset? ValidTo,
-        string? ActorNumber,
-        string? ActorName,
-        DateTimeOffset? FullFlexDate);
+    public sealed record GetGridAreaAuditLogEntriesResponse(IEnumerable<GridAreaAuditLogEntryDto> GridAreaAuditLogEntries);
 }