+++ source/Energinet.DataHub.MarketParticipant.Client/documents/release-notes/release-notes.md
@@ -1,5 +1,13 @@
 # Energinet.DataHub.MarketParticipant.Client Release notes
 
+## Version 2.6.0
+
+- User display name on audit log entry
+
+## Version 2.5.2
+
+- API for retrieving audit log entries for a grid area.
+
 ## Version 2.5.1
 
 - Do not handle unexpected exceptions.